+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Added: `Action.next()` as a substitute to existing `MachineState.nextAction`.
     The regular `Store` class now also supports behavior of `StateMachine`,
     which makes `StateMachine` obsolete (will be removed before stable release).
+- Deprecated: `StateMachine` and related classes
+    (`MachineState`, `StateMachineBuilder`). Regular `Store` class can now be
+    used as a state machine.
 
 ## 1.0.0-dev.1.0